Abstracted out the Solution class
WARNING: still contains errors and is untested

Made it so the solution space is divided into Counting and Matching solutions. This will make it easier to add new kinds of solutions and deal with new kinds of problems

Matching solutions are those which there can be a match between an element in one solution and an element of annother solution of the same type.
Counting solutions are ones in which it is more about counting the elements or keeping a count associated with elements.